restructured userpage
