+++ package.json
@@ -20,6 +20,7 @@
   "dependencies": {
     "@auth0/nextjs-auth0": "^0.11.0",
     "@date-io/date-fns": "^2.5.0",
+    "@mailchimp/mailchimp_marketing": "^3.0.24",
     "@material-ui/core": "^4.9.5",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "^4.0.0-alpha.45",
