--- src/components/Userview/index.js
@@ -1,13 +1,0 @@
-import ProfileData from './ProfileData';
-import ProfileHeader from './ProfileHeader';
-import ProfileBookList from './ProfileBookList';
-import ProfileBookListItem from './ProfileBookListItem';
-import ProfileImage from './ProfileImage';
-
-export {
-  ProfileData,
-  ProfileHeader,
-  ProfileImage,
-  ProfileBookList,
-  ProfileBookListItem,
-};
