+++ src/pages/UserPage/components/ProfileBookListItem.jsx
@@ -1,7 +1,7 @@
 import { makeStyles } from '@material-ui/core/styles';
 import { Typography, Link } from '@material-ui/core';
-import urlGenerator from '../../helpers/urlGenerator';
-import generateAuthorString from '../../helpers/generateAuthorString'
+import urlGenerator from '../../../helpers/urlGenerator';
+import generateAuthorString from '../../../helpers/generateAuthorString'
 
 const useStyles = makeStyles((theme) => ({
   listItem: {
