+++ src/pages/UserPage/index.tsx
@@ -1,0 +1,13 @@
+import ProfileData from './components/ProfileData';
+import ProfileHeader from './components/ProfileHeader';
+import ProfileBookList from './components/ProfileBookList';
+import ProfileBookListItem from './components/ProfileBookListItem';
+import ProfileImage from './components/ProfileImage';
+
+export {
+  ProfileData,
+  ProfileHeader,
+  ProfileImage,
+  ProfileBookList,
+  ProfileBookListItem,
+};
